Refactor: components에 common 폴더 추가
