+++ src/pages/Main.tsx
@@ -15,7 +15,7 @@ import { useMediaQuery } from 'react-responsive'
 import Board from '@src/components/Board'
 import useInfinityScroll from '@src/hooks/useInfinityScroll'
 import { categoryNamesList } from '@src/constants/options'
-import Loading from '@src/components/loading'
+import Loading from '@src/components/common/loading'
 
 const Main = () => {
   const isMobile = useMediaQuery({
